Add data integrity on nameID, email, upn (#5107)
* Adds unique constraints to entities

Ensures data integrity by adding unique constraints to specific fields
in multiple entities, preventing duplicate entries. This includes
'nameID' in several entities and 'accountUpn' and 'email' in the User
entity. A new migration is added to create the unique indices.

* Ensures accountUpn is synchronized with email

Updates existing user records to ensure that the `accountUpn` field matches the `email` field.

Also adds a unique index to the `nameID` column in the `user` table to enforce uniqueness.