Adds subdomain whitelisting to innovation hub (#5125)
* Adds subdomain whitelisting to innovation hub

Ensures that specific subdomains are excluded from being treated as innovation hubs.

This prevents unintended routing and ensures that designated subdomains, such as 'identity', are not intercepted by the innovation hub logic. The whitelisted subdomains are configurable through the application's configuration.

* Configures innovation hub whitelisted subdomains

Updates the configuration to allow specifying a comma-separated list of subdomains that should not be treated as innovation hubs.
Previously, the whitelisted subdomains were hardcoded. Now the list is configurable via the alkemio.yml file and parsed as comma separated string.