LicenseCounts (#5222)
* updated counting of space entitlements

* Refactors space entitlement counting logic

Simplifies space entitlement counting by introducing a priority-based system.

This ensures that the highest priority enabled entitlement is correctly
identified for each space, and improves the accuracy of entitlement usage
calculations. Uses a constant to define entitlement precedence.

Adds comprehensive unit tests to validate the new logic.

* Refactors EntityManager mock in tests

Updates the test setup to directly mock the EntityManager using NestJS's `getEntityManagerToken`.

This change improves test isolation and avoids relying on implementation details of the service. It also allows to mock the `findOne` method directly on the EntityManager instance.

---------

Co-authored-by: Neil Smyth <neil@thesmyths.eu>