Improves contribution reporting with user details (#5533)
Enhances contribution reporting by resolving and including user details such as ID and email when a contributor joins a space. This ensures more accurate and informative contribution tracking.
The author details are resolved based on the type of contributor (User, Organization, or VirtualContributor).
Also, fixes a typo in `contribution-reporter.service.ts` where the space ID was being incorrectly set.

Co-authored-by: Svetoslav Petkov <svetoslav@alkem.io>