Reduces log verbosity for collaboration checks (#5534)
* Reduces log verbosity for collaboration checks

Decreases the log level from warn to verbose when a user is unable to read a collaboration.

This change reduces the noise in the logs by only logging these events at a more detailed verbosity level, as they are not necessarily indicative of a problem.

* Remove logging for unreadable collaboration access

Removed verbose logging for unreadable collaborations.

---------

Co-authored-by: Svetoslav Petkov <svetoslav@alkem.io>